Converted everything to daemon, and incrimental map reduce

+++ matchProcessor/GetAverageItemBuild/main.js
@@ -1,8 +1,8 @@
 
 var mongodb = require('mongodb');
+ObjectId = require('mongodb').ObjectID;
 
-
-var map = function(){
+var mapFunction = function(){
 
     // ANything with group containing "flasks", "relicBase" "ward" "potion"
     var skippedItems = [
@@ -55,7 +55,8 @@ var map = function(){
     emit(victimKey, victimRetVal);
 };
 
-var reduce = function (key, resultsList){
+var reduceFunction = function (key, resultsList){
+
     var retval = {};
     resultsList.forEach(function (itemBuildMap){
     // go through each type of build within the build
@@ -72,33 +73,82 @@ var reduce = function (key, resultsList){
 
 function finalize (key,value){
     var res=key.split("-");
-
+    var retval = {"ChampId":res[0],"MinuteMark":res[1]};
     var sortable = [];
     for(var build in value){
+        retval[build] = value[build];
         sortable.push([build,value[build]]);
 
     }
     sortable.sort(function(a,b){return b[1] - a[1]});
-    var topBuild = sortable[0];
-    return({"ChampId":res[0],"MinuteMark":res[1],"TopBuild":topBuild});
+    retval['TopBuild'] = sortable[0];
+
+    return(retval);
 
 }
 
 
 
-db.AverageItemBuildPerMinute.drop();
-db.totalKillCollection.mapReduce(map,reduce,{out:'AverageItemBuildPerMinute',finalize:finalize});
-
-
-
 var db;
 mongodb.connect ('mongodb://mongo:27017/urfday',function(err,db){
+
     if  (err){return console.dir(err);}
-    console.log("We are connected to the DB");
-    db.collection('AverageItemBuildPerMinute').drop();
-    db.collection('totalKillCollection').mapReduce(map,reduce,{out:'AverageItemBuildPerMinute',finalize:finalize});
-    console.log("Map reduce complete");
-    return;
+
+
+
+    function setNextLoop(){
+        setTimeout(function(){
+            var pointerCollection = db.collection('MapReducePointers');
+            var startPoint;
+
+            pointerCollection.find({Process:'AvgItemBuild'}).limit(1).toArray(function(err, docs) {
+                if  (err){return console.dir(err);}
+
+                if(docs.length){
+                    console.log("Map reduce previously run. Incrementing ...");
+                    startPoint = docs[0]['Location'];
+                    console.log("Picking up where we stopped at item " + startPoint);
+                }
+                else{
+                    db.collection('AverageItemBuildPerMinute').drop();
+                    console.log("Never Run before, Starting The DB fresh");
+                    startPoint = 0;
+                }
+
+                pointerCollection.insert({Process:'AvgItemBuild',KillsToOdds:startPoint},function(err,result){
+                    if  (err){return console.dir(err);}
+
+                    db.collection('totalKillCollection').find().sort({_id:-1}).limit(1).toArray(function(err, lastItem) {
+                        if  (err){return console.dir(err);}
+
+                        var LastItemAdded = lastItem[0]['_id'];
+
+                        console.log("Looking for items between " + startPoint + " and " + LastItemAdded);
+
+                        var startPointObject = new ObjectId(startPoint);
+                        var LastItemAddedObject = new ObjectId(LastItemAdded)
+                         //query = {_id:{$gt:startPointObject}};
+                        query = { $and: [{ _id: {$gt:startPointObject}},{_id:{$lte:LastItemAddedObject}}]};
+
+                        console.log("Map reduce Started");
+                        db.collection('totalKillCollection').mapReduce(mapFunction,reduceFunction,{out:{reduce:'AverageItemBuildPerMinute'},finalize:finalize,query:query,verbose:true},function(err,collection,stats){//finalize:finalize
+                            if(err){return console.dir(err);}
+                            console.log("Map-Reduce completed")
+                            console.log(stats)
+                            db.collection('MapReducePointers').findOneAndUpdate({Process:'AvgItemBuild'},{Process:'AvgItemBuild', Location:LastItemAdded},function(err,db){
+                                if  (err){return console.dir(err);}
+                                console.log("Updated Pointer of incremental map reduce to " + LastItemAdded)
+                                setNextLoop()
+                            });
+                        });
+                    });
+                });
+            });
+        },300000);
+    }
+
+    setNextLoop();
+
 
 });
